# Simplify tail index invariant