@christianrpetrin 2018-11-29
> Benchmark note: I've noticed that getTestValue allocates a new
> value, which means that a lot of the benchmarks are essentially
> just benchmarking the garbage collector. I'd suggest changing
> getTestValue so that it doesn't allocate (it could just return nil)
> so we're measuring more of the Deque operations themselves.
> Also, by not using the Direct benchmarks, we'll be seeing a significant
> amount of overhead from the indirect function calls.
> How about moving all the comparison benchmarks into another package
> (or even another repository, losing the additional dependencies)
> and have the benchmarks in the deque package call Deque directly?

I love your thinking! It's funny because yesterday just before going to bed I had a similar thought: it would be better to move all bench tests (except the direct ones) to another repo, so we could remove all dependencies to the external queues. It would make the repo considerably cleaner and leaner.

I'll create a new repo and move all bench tests there. Will also do the change on the direct tests. ETA: tomorrow EOD.