@rogpeppe 2018-11-29
> I love your thinking! It's funny because yesterday just before going to bed I had a similar thought: it would be better to move all bench tests (except the direct ones) to another repo, so we could remove all dependencies to the external queues. It would make the repo considerably cleaner and leaner.

I did some stuff on this, but didn't complete it. See #6.